# Changing subtile selection in the try_centroid_placement of initial_placement