--- a/vtr_flow/tasks/regression_tests/vtr_reg_strong_odin/strong_sdc/config/config.txt
+++ b/vtr_flow/tasks/regression_tests/vtr_reg_strong_odin/strong_sdc/config/config.txt
@@ -24,7 +24,7 @@ qor_parse_file=qor_standard.txt
 pass_requirements_file=pass_requirements.txt
 
 # Script parameters
-script_params_common = -starting_stage vpr
+script_params_common = -starting_stage vpr --seed 6
 script_params_list_add = -sdc_file sdc/samples/A.sdc
 script_params_list_add = -sdc_file sdc/samples/B.sdc
 script_params_list_add = -sdc_file sdc/samples/C.sdc