<!--- Provide a general summary of your changes in the Title above -->
Summary: changed the subtile selection in the initial_placement to prevent falling into random macro placement while we have available subtiles in the selected location.
#### Description
<!--- Describe your changes in detail -->

There was discrepancy in the MCNC benchmark results while reconstructing from flat placement. The initial results can be seen in the "[APPack] Added Max Displacement Metric #2876". The discrepancy was that even though we were able to reconstruct the clusters with zero error, there were displacements in atoms. 

I found that the displacement occurs when we fall into try_place_macro_randomly even though we have available subtiles in the desired location. This seems to be caused by trying to place the block into a randomly selected compatible subtile and if it is occupied falling into try_place_macro_randomly. My change suggests to select the subtile in the desired location from the unused compatible subtitles, falling back into try_place_macro_randomly if we do not have unused compatible subtile in the desired location.

An example of the results indicating discrepancy also given for the ALU4 circuit in the tables below before and after the change.

Before change:

| iter_num | pack_time(s) | device_util | total_wl | CPD(ns) | percent_cluster_errors | percent_atoms_displaced | average_atom_displacement | max_atom_displacement |
|----------|--------------|-------------|----------|---------|------------------------|-------------------------|--------------------------|-----------------------|
| 0        | 0.70         | 0.67        | 6398     | 4.92578 | -1                     | -1                      | -1                       | -1                    |
| 1        | 0.56         | 0.68        | 6413     | 5.06346 | 0.227723               | 0.261242                | 1.406852                 | 12.000000             |
| 2        | 0.62         | 0.68        | 6276     | 5.0679  | 0.039604               | 0.061028                | 0.265525                 | 10.000000             |
| 3        | 0.58         | 0.68        | 6256     | 5.00301 | 0.000000               | 0.004283                | 0.033191                 | 11.000000             |
| 4        | 0.57         | 0.68        | 6284     | 5.12653 | 0.000000               | 0.001071                | 0.002141                 | 2.000000              |
| 5        | 0.59         | 0.68        | 6277     | 5.0527  | 0.000000               | 0.002141                | 0.020343                 | 10.000000             |
| 6        | 0.57         | 0.68        | 6345     | 5.00962 | 0.000000               | 0.002141                | 0.010707                 | 7.000000              |
| 7        | 0.56         | 0.68        | 6318     | 4.96222 | 0.000000               | 0.000000                | 0.000000                 | 0.000000              |
| 8        | 0.55         | 0.68        | 6245     | 5.06431 | 0.000000               | 0.001071                | 0.011777                 | 11.000000             |
| 9        | 0.55         | 0.68        | 6311     | 4.95586 | 0.000000               | 0.000000                | 0.000000                 | 0.000000              |
| 10       | 0.58         | 0.68        | 6339     | 4.96665 | 0.000000               | 0.000000                | 0.000000                 | 0.000000              |


After change:

| iter_num | pack_time(s) | device_util | total_wl | CPD(ns) | percent_cluster_errors | percent_atoms_displaced | average_atom_displacement | max_atom_displacement |
|----------|--------------|-------------|----------|---------|------------------------|-------------------------|--------------------------|-----------------------|
| 0        | 0.64         | 0.67        | 6271     | 4.93428 | -1                     | -1                      | -1                       | -1                    |
| 1        | 0.57         | 0.68        | 6362     | 4.96405 | 0.227723               | 0.225910                | 1.207709                 | 13.000000             |
| 2        | 0.57         | 0.68        | 6371     | 5.08332 | 0.039604               | 0.130621                | 0.509636                 | 11.000000             |
| 3        | 0.54         | 0.68        | 6491     | 4.98076 | 0.000000               | 0.000000                | 0.000000                 | 0.000000              |
| 4        | 0.54         | 0.68        | 6446     | 4.98408 | 0.000000               | 0.000000                | 0.000000                 | 0.000000              |
| 5        | 0.54         | 0.68        | 6391     | 4.97744 | 0.000000               | 0.000000                | 0.000000                 | 0.000000              |
| 6        | 0.59         | 0.68        | 6357     | 4.986   | 0.000000               | 0.000000                | 0.000000                 | 0.000000              |
| 7        | 0.55         | 0.68        | 6344     | 5.04787 | 0.000000               | 0.000000                | 0.000000                 | 0.000000              |
| 8        | 0.57         | 0.68        | 6282     | 5.04263 | 0.000000               | 0.000000                | 0.000000                 | 0.000000              |
| 9        | 0.55         | 0.68        | 6287     | 5.05193 | 0.000000               | 0.000000                | 0.000000                 | 0.000000              |
| 10       | 0.54         | 0.68        | 6317     | 4.99501 | 0.000000               | 0.000000                | 0.000000                 | 0.000000              | 


#### Related Issue
<!--- Pull requests should be related to open issues -->
<!--- If suggesting a new feature or change, please discuss it in an issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps to reproduce -->
<!--- Please link to the issue here: -->
This is not an issue but related to the results mentioned in "[APPack] Added Flat Placement Reconstruction #2870".

#### Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
To investigate the discrepancy between cluster reconstruction error and atom displacement, especially in the MCNC benchmarks.

#### How Has This Been Tested?
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran to -->
<!--- see how your change affects other areas of the code, etc. -->
Basic and strong tests.
#### Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [x] Bug fix (change which fixes an issue)
- [ ] New feature (change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)

#### Checklist:
<!--- Go over all the following points, and put an `x` in all the boxes that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [ ] My change requires a change to the documentation
- [x] I have updated the documentation accordingly
- [ ] I have added tests to cover my changes
- [ ] All new and existing tests passed