@AlexandreSinger 2025-02-21
I have launched the Nigthly Tests for this branch:
https://github.com/verilog-to-routing/vtr-verilog-to-routing/actions/runs/13467086924